feat : model/domain 패키지로 Lotto 클래스 위치 변경 및 세부 메소드 구현 - validate() : lotto의 규격에 맞지 않은 값들이 들어왔을 때 ERROR 발생 - getNumbers() : numbers 반환
